backend/API/gmo_backend/requirements.txt
aniso8601==7.0.0
asgiref==3.2.10
attrs==20.3.0
certifi==2020.11.8
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.6.0
django-graphql-geojson==0.1.4
drf-yasg==1.20.0
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0

appdirs==1.4.4
argon2-cffi==20.1.0
async-generator==1.10
backcall==0.2.0
bcrypt==3.1.7
bleach==3.2.1
blinker==1.4
cached-property==1.5.1
cffi==1.14.4
cryptography==2.8
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
distro==1.4.0
Django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.12.1
docopt==0.6.2
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
GDAL==3.0.4
httplib2==0.14.0
importlib-metadata==1.5.0
ipykernel==5.4.0
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
json5==0.9.5
jsonschema==3.2.0
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
netifaces==0.10.4
notebook==6.1.5
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
openapi-codec==1.3.2
packaging==20.7
pandocfilters==1.4.3
paramiko==2.6.0
parso==0.7.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.6.1
psycopg2==2.8.6
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.3
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.15.5
python-dateutil==2.7.3
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==20.0.0
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.1
terminado==0.9.1
testpath==0.4.4
texttable==1.6.2
tornado==6.1
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.8
whitenoise==5.2.0
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.53.0
zipp==1.0.0
drf-yasg==1.20.0
django-model-mutations
django-graphql-jwt==0.3.0
django-graphql-auth